requirements.txt
kafka-python==2.0.2
numpy
scikit-learn
boto3
cachetools==5.3.0
